[level 2] Title: 점 찍기, Time: 11.84 ms, Memory: 3.62 MB -BaekjoonHub

+++ 프로그래머스/2/140107. 점 찍기/점 찍기.cpp
@@ -7,7 +7,7 @@ ll solution(int k, int d) {
     ll answer = 0;
     ll dist = (ll)d * d;
     for (ll i=0; i <= d; i += k) {
-        long long rootY = (long long)sqrt(dist - i * i);
+        long long rootY = sqrt(dist - i * i);
         answer += (rootY / k + 1);
     }
     return answer;
